pyproject.toml
[project]
name = "fastapistore"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.43.0,<0.44.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "sqlalchemy[asyncio] (>=2.0.49,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "fastapi-debug-toolbar (>=0.6.3,<0.7.0)",
    "pwdlib (>=0.3.0,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.20.0"
black = "^26.3.1"
isort = "^8.0.1"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
    | \.migrations
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
